backtrader_system/requirements.txt
# Backtrader ML Trading System Requirements

# Core dependencies
backtrader>=1.9.76.123
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
yfinance>=0.1.70
matplotlib>=3.3.0
PyYAML>=5.4.0

# Optional for enhanced features
streamlit>=1.10.0  # For dashboard
plotly>=5.0.0      # For interactive plots
alpaca-trade-api>=2.0.0  # For live trading

# Development
pytest>=6.0.0
black>=21.0.0
flake8>=3.9.0
